% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_condition_functions.R
\name{and_filter_condition}
\alias{and_filter_condition}
\title{Combine two conditions with AND (&&)}
\usage{
and_filter_condition(fcondition1, fcondition2)
}
\arguments{
\item{fcondition1}{First condition list}

\item{fcondition2}{Second condition list}
}
\value{
Combined condition list with merged arrays/files
}
\description{
Combine two conditions with AND (&&)
}
\keyword{internal}
