% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk.R
\name{as_block}
\alias{as_block}
\title{Wrap code in awk block syntax { ... }}
\usage{
as_block(code)
}
\arguments{
\item{code}{Awk code string to wrap}
}
\value{
Code wrapped in braces with proper indentation
}
\description{
Wrap code in awk block syntax { ... }
}
\keyword{internal}
