% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_class.R
\name{as_filter_condition}
\alias{as_filter_condition}
\alias{as_filter_condition.default}
\alias{as_filter_condition.quosure}
\title{Coerce an object to a filter_condition}
\usage{
as_filter_condition(x, ...)

\method{as_filter_condition}{default}(x, ...)

\method{as_filter_condition}{quosure}(x, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{...}{Additional attributes to set}
}
\value{
Object with filter_condition class added
}
\description{
Adds the filter_condition class to an object while preserving existing
classes.
}
\keyword{internal}
