% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_positions.R
\name{as_genomic_positions}
\alias{as_genomic_positions}
\title{Coerce to genomic_positions}
\usage{
as_genomic_positions(genomic_positions_dt, ...)
}
\arguments{
\item{genomic_positions_dt}{A data.table or genomic_regions object}

\item{...}{Additional arguments passed to as_genomic_regions}
}
\value{
A genomic_positions object
}
\description{
Coerce to genomic_positions
}
\keyword{internal}
