% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk.R
\name{awk_split_column}
\alias{awk_split_column}
\title{Generate awk split() function call}
\usage{
awk_split_column(bash_index, array_name, delimiter)
}
\arguments{
\item{bash_index}{The awk column reference (e.g., "$1")}

\item{array_name}{Name for the awk array to receive split values}

\item{delimiter}{Delimiter character for splitting}
}
\value{
Character string like `split($1, arr, ";")`
}
\description{
Creates awk code to split a column value into an array using a delimiter.
}
\keyword{internal}
