% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{build<-}
\alias{build<-}
\alias{build<-.default}
\alias{build<-.data.table}
\alias{build<-.list}
\alias{build<-.file_interface}
\title{Set genome build on an object}
\usage{
build(x) <- value

\method{build}{default}(x) <- value

\method{build}{data.table}(x) <- value

\method{build}{list}(x) <- value

\method{build}{file_interface}(x) <- value
}
\arguments{
\item{x}{Object to modify}

\item{value}{Build string to set (e.g., "b37", "b38")}
}
\value{
Modified object
}
\description{
Replacement function for setting the genome build attribute.
Uses data.table::setattr for data.tables (modifies by reference).
}
