% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_codegen.R
\name{build_awk_file_args}
\alias{build_awk_file_args}
\title{Build awk file arguments string}
\usage{
build_awk_file_args(finterface, branches, array_loads, use_command_line_fs)
}
\arguments{
\item{finterface}{File interface}

\item{branches}{Branch data.table with process_substitution}

\item{array_loads}{Array-load data.table with temp_file}

\item{use_command_line_fs}{Whether to set FS on command line}
}
\value{
File arguments string for awk command
}
\description{
Assembles file arguments: process substitutions (tabix), array-load temp
files (\%in\% arrays), and main data file with appropriate separators.
}
\keyword{internal}
