% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_build.R
\name{build_can_be_inferred}
\alias{build_can_be_inferred}
\title{Check if build can be inferred from summary statistics}
\usage{
build_can_be_inferred(summary_stats)
}
\arguments{
\item{summary_stats}{data.table to check}
}
\value{
Logical indicating whether required columns exist
}
\description{
Check if build can be inferred from summary statistics
}
\keyword{internal}
