% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_codegen.R
\name{build_file_read_cmd}
\alias{build_file_read_cmd}
\title{Build shell command to read file (cat or zcat)}
\usage{
build_file_read_cmd(finterface)
}
\arguments{
\item{finterface}{File interface with filename and gzipped flag}
}
\value{
Shell command string: "cat file" or "zcat file.gz"
}
\description{
Build shell command to read file (cat or zcat)
}
\keyword{internal}
