% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_codegen.R
\name{build_line_limit_code}
\alias{build_line_limit_code}
\title{Build awk code for line limit enforcement}
\usage{
build_line_limit_code(nlines)
}
\arguments{
\item{nlines}{Maximum number of lines to output}
}
\value{
Awk if statement for line limiting, or NULL if no limit
}
\description{
Build awk code for line limit enforcement
}
\keyword{internal}
