% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk.R
\name{build_pipeline_cmd}
\alias{build_pipeline_cmd}
\title{Build pipeline command (file | awk)}
\usage{
build_pipeline_cmd(file_cmd, awk_script)
}
\arguments{
\item{file_cmd}{Shell command to read file}

\item{awk_script}{Awk script/command}
}
\value{
Piped command string
}
\description{
Build pipeline command (file | awk)
}
\keyword{internal}
