% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk.R
\name{build_trim_prefix_code}
\alias{build_trim_prefix_code}
\title{Build awk code to remove line prefix}
\usage{
build_trim_prefix_code(trim_prefix)
}
\arguments{
\item{trim_prefix}{Prefix to remove from each line}
}
\value{
Awk gsub() call to remove prefix, or NULL if no prefix
}
\description{
Build awk code to remove line prefix
}
\keyword{internal}
