% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_internal.R
\name{check_quotes_to_write}
\alias{check_quotes_to_write}
\title{Format values for writing to temp file}
\usage{
check_quotes_to_write(values, quoted)
}
\arguments{
\item{values}{Vector of values}

\item{quoted}{TRUE if column values are quoted in the file}
}
\value{
Formatted values for temp file writing
}
\description{
Prepares values for writing to temp files (used by %in% operator).
For quoted columns, ensures values are properly quoted to match
file contents.
}
\keyword{internal}
