% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_info.R
\name{check_single_column_prefix}
\alias{check_single_column_prefix}
\title{Check if all column values have a prefix}
\usage{
check_single_column_prefix(prefixes, column_data)
}
\arguments{
\item{prefixes}{Vector of possible prefixes to check}

\item{column_data}{Vector of values from the column}
}
\value{
First matching prefix, or NULL if none match
}
\description{
Check if all column values have a prefix
}
\keyword{internal}
