% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_ops.R
\name{chr_has_mixed_na}
\alias{chr_has_mixed_na}
\title{Check if genomic_regions has both NA and non-NA chromosomes}
\usage{
chr_has_mixed_na(gregions)
}
\arguments{
\item{gregions}{A genomic_regions object}
}
\value{
TRUE if some rows have NA chr and others don't
}
\description{
Check if genomic_regions has both NA and non-NA chromosomes
}
\keyword{internal}
