% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_info.R
\name{column_names}
\alias{column_names}
\title{Get column names from file interface}
\usage{
column_names(finterface, original = FALSE)
}
\arguments{
\item{finterface}{File interface object}

\item{original}{If TRUE, return original input names; else return
standardized names}
}
\value{
Character vector of column names
}
\description{
Get column names from file interface
}
\keyword{internal}
