% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_class.R
\name{combine_by_embedded_build}
\alias{combine_by_embedded_build}
\title{Split a mixed-build variant_id table by build and recombine}
\usage{
combine_by_embedded_build(x, parsed)
}
\description{
Split a mixed-build variant_id table by build and recombine
}
\keyword{internal}
