% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_ops.R
\name{compute_intersection}
\alias{compute_intersection}
\title{Compute intersection using foverlaps}
\usage{
compute_intersection(e1, e2, build)
}
\arguments{
\item{e1, e2}{Prepared genomic_regions (sentinel values, keyed)}

\item{build}{Build to assign to result}
}
\value{
genomic_regions with overlapping portions
}
\description{
Compute intersection using foverlaps
}
\keyword{internal}
