% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_class.R
\name{copy_genomic_regions}
\alias{copy_genomic_regions}
\title{Deep copy genomic_regions}
\usage{
copy_genomic_regions(gregions)
}
\arguments{
\item{gregions}{A genomic_regions object}
}
\value{
A copy with class preserved
}
\description{
Deep copy genomic_regions
}
\keyword{internal}
