% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_interface.R
\name{detect_comment_prefix}
\alias{detect_comment_prefix}
\title{Detect comment prefix from a file}
\usage{
detect_comment_prefix(finterface)
}
\arguments{
\item{finterface}{A file_interface object}
}
\value{
Character string with the comment prefix pattern, or NULL
}
\description{
Identifies the prefix of commented lines at the beginning of a file to drop
them when parsing the file.
}
\keyword{internal}
