% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{drop_chr_prefix}
\alias{drop_chr_prefix}
\title{Remove "chr" prefix from chromosome names}
\usage{
drop_chr_prefix(chr)
}
\arguments{
\item{chr}{Character vector of chromosome names}
}
\value{
Chromosome names without "chr" prefix
}
\description{
Remove "chr" prefix from chromosome names
}
\keyword{internal}
