% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_class.R
\name{empty_filter_condition}
\alias{empty_filter_condition}
\title{Create an empty filter condition}
\usage{
empty_filter_condition(
  build = NULL,
  genomic_regions = full_genomic_regions(build = build),
  finterface_env = NULL
)
}
\arguments{
\item{build}{Optional genome build (b36, b37, b38)}

\item{genomic_regions}{Genomic regions to associate with the condition}

\item{finterface_env}{Environment containing the file_interface}
}
\value{
An empty filter_condition object
}
\description{
Creates a filter_condition with no conditions (no filtering, matches all
rows).
}
\keyword{internal}
