% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_class.R
\name{empty_genomic_regions}
\alias{empty_genomic_regions}
\title{Create empty genomic_regions (no regions)}
\usage{
empty_genomic_regions(build = NULL)
}
\arguments{
\item{build}{Genome build identifier}
}
\value{
Empty genomic_regions (0 rows)
}
\description{
Create empty genomic_regions (no regions)
}
\keyword{internal}
