% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_condition_functions.R
\name{eq_filter_condition}
\alias{eq_filter_condition}
\title{Build awk equality condition}
\usage{
eq_filter_condition(var1, var2)
}
\arguments{
\item{var1}{Left operand (column ref or value)}

\item{var2}{Right operand (column ref or value)}
}
\value{
List with `condition` element: "var1 == var2"
}
\description{
Build awk equality condition
}
\keyword{internal}
