% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRead-package.R
\docType{package}
\name{filterRead-package}
\alias{filterRead}
\alias{filterRead-package}
\title{filterRead: Pre-filters files using awk for faster fread}
\description{
Creates an easy-to-use interface for loading tab/comma-separated value files efficiently. Conditions for filtering can be specified in R-style and are translated to awk command line before passing on to data.table::fread for loading.
}
\keyword{internal}
