% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_utils.R
\name{genomic_regions<-}
\alias{genomic_regions<-}
\title{Set genomic regions on a filter condition}
\usage{
genomic_regions(fcondition) <- value
}
\arguments{
\item{fcondition}{A filter_condition object}

\item{value}{A genomic_regions object to set}
}
\value{
The modified filter_condition
}
\description{
Set genomic regions on a filter condition
}
\keyword{internal}
