% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_metadata.R
\name{get_base_column_info}
\alias{get_base_column_info}
\title{Create base column info from file headers}
\usage{
get_base_column_info(
  finterface,
  standard_names_dt = summary_stats_standard_names_dt
)
}
\arguments{
\item{finterface}{File interface object}

\item{standard_names_dt}{Data.table mapping input names to standard names}
}
\value{
data.table with base column metadata
}
\description{
Initializes column_info data.table with input names, indices, and
joins with standard name mappings.
}
\keyword{internal}
