% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_build.R
\name{get_build_from_file_interface}
\alias{get_build_from_file_interface}
\title{Get or infer genome build from a file interface}
\usage{
get_build_from_file_interface(finterface, nsnps = 10000)
}
\arguments{
\item{finterface}{A file_interface object}

\item{nsnps}{Number of SNPs to use for build inference}
}
\value{
The genome build string (b36, b37, b38) or NA_character_
}
\description{
Returns the build if already set, otherwise infers it from the file data.
}
\keyword{internal}
