% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{get_chain_downloads}
\alias{get_chain_downloads}
\title{Get chain file download URLs and filenames}
\usage{
get_chain_downloads(from, to)
}
\arguments{
\item{from}{Source build (b36, b37, b38, or synonyms like hg38)}

\item{to}{Target build (b36, b37, b38, or synonyms like hg38)}
}
\value{
List with url, filename, and uncompressed filename
}
\description{
Get chain file download URLs and filenames
}
\keyword{internal}
