% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{get_chain_filename}
\alias{get_chain_filename}
\title{Get chain filename for a build conversion}
\usage{
get_chain_filename(from, to)
}
\arguments{
\item{from}{Source build (b36, b37, b38, or synonyms like hg38)}

\item{to}{Target build (b36, b37, b38, or synonyms like hg38)}
}
\value{
Chain filename (uncompressed)
}
\description{
Get chain filename for a build conversion
}
\keyword{internal}
