% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_chain_path}
\alias{get_chain_path}
\title{Get chain files path}
\usage{
get_chain_path(warn = TRUE)
}
\arguments{
\item{warn}{Logical, whether to throw a warning if the path is not configured}
}
\description{
Get chain files path
}
