% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{get_chain_rds_filename}
\alias{get_chain_rds_filename}
\title{Get RDS cache filename for a chain file}
\usage{
get_chain_rds_filename(from, to)
}
\arguments{
\item{from}{Source build (b36, b37, b38, or synonyms like hg38)}

\item{to}{Target build (b36, b37, b38, or synonyms like hg38)}
}
\value{
RDS cache filename using internal build names
}
\description{
Get RDS cache filename for a chain file
}
\keyword{internal}
