% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_metadata.R
\name{get_column_info}
\alias{get_column_info}
\title{Build complete column info for a file interface}
\usage{
get_column_info(
  finterface,
  standard_names_dt = summary_stats_standard_names_dt,
  nrows_to_check = 500
)
}
\arguments{
\item{finterface}{File interface object}

\item{standard_names_dt}{Data.table mapping input names to standard names}

\item{nrows_to_check}{Number of rows to sample for pattern detection}
}
\value{
data.table with complete column metadata
}
\description{
Main orchestrator that detects and configures all column metadata by
calling sub-functions for each type of detection.
}
\keyword{internal}
