% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp.R
\name{get_dbsnp_downloads}
\alias{get_dbsnp_downloads}
\title{Get all dbSNP files to download}
\usage{
get_dbsnp_downloads(build = c("b37", "b38"), type = c("common", "all"))
}
\arguments{
\item{build}{Character vector of builds ("b37", "b38")}

\item{type}{Type of SNP reference set (common = .01 <= MAF)
("common" or "all")}
}
\description{
Get all dbSNP files to download
}
\keyword{internal}
