% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp.R
\name{get_dbsnp_filename}
\alias{get_dbsnp_filename}
\title{Get dbSNP local filename}
\usage{
get_dbsnp_filename(build = c("b37", "b38"), type = c("common", "all"))
}
\arguments{
\item{build}{Build version ("b37" or "b38")}

\item{type}{Type of SNP reference set (common = .01 <= MAF)
("common" or "all")}
}
\description{
Get dbSNP local filename
}
\keyword{internal}
