% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_dbsnp_path}
\alias{get_dbsnp_path}
\title{Get dbSNP reference file path}
\usage{
get_dbsnp_path(warn = TRUE)
}
\arguments{
\item{warn}{Logical, whether to throw a warning if the path does not exist}
}
\description{
Get dbSNP reference file path
}
