% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_matching.R
\name{get_dbsnp_url}
\alias{get_dbsnp_url}
\title{Get dbSNP file URL}
\usage{
get_dbsnp_url(build = c("b37", "b38"), type = c("common", "all"))
}
\arguments{
\item{build}{Build version ("b37" or "b38")}

\item{type}{Type of SNP reference set (common = .01 <= MAF)
("common" or "all")}
}
\description{
Get dbSNP file URL
}
\keyword{internal}
