% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_class.R
\name{get_file_interface}
\alias{get_file_interface}
\title{Get the file interface from a filter condition}
\usage{
get_file_interface(fcondition)
}
\arguments{
\item{fcondition}{A filter_condition object}
}
\value{
The file_interface object, or NULL if not set
}
\description{
Extracts the file interface object stored in the filter condition's
finterface_env attribute.
}
\keyword{internal}
