% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_interface.R
\name{get_file_separator}
\alias{get_file_separator}
\title{Detect the field separator of a file}
\usage{
get_file_separator(finterface)
}
\arguments{
\item{finterface}{A file_interface object}
}
\value{
Character string containing the detected separator
}
\description{
Uses data.table's fread to detect the most likely field separator
by analyzing the first 100 lines of the file.
}
\keyword{internal}
