% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_build.R
\name{get_tabix_matches}
\alias{get_tabix_matches}
\title{Count matching positions in a reference file using tabix}
\usage{
get_tabix_matches(summary_stats, ref_filename)
}
\arguments{
\item{summary_stats}{data.table with chr, pos, ref, alt columns}

\item{ref_filename}{Path to tabix-indexed reference file}
}
\value{
Integer count of matching variants
}
\description{
Count matching positions in a reference file using tabix
}
\keyword{internal}
