% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_matching.R
\name{get_tabix_process_substitution}
\alias{get_tabix_process_substitution}
\title{Build a tabix process substitution command}
\usage{
get_tabix_process_substitution(chr, start, end, dbsnp_filename)
}
\arguments{
\item{chr}{Chromosome(s)}

\item{start}{Start position(s)}

\item{end}{End position(s)}

\item{dbsnp_filename}{Path to tabix-indexed dbSNP file}
}
\value{
Character string with process substitution command
}
\description{
Creates a bash process substitution that queries tabix for genomic regions.
}
\keyword{internal}
