% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_build.R
\name{get_tabix_query}
\alias{get_tabix_query}
\title{Build a tabix query command for summary statistics}
\usage{
get_tabix_query(summary_stats, ref_filename)
}
\arguments{
\item{summary_stats}{data.table with chr and pos columns}

\item{ref_filename}{Path to the tabix-indexed reference file}
}
\value{
Character string containing the tabix command
}
\description{
Build a tabix query command for summary statistics
}
\keyword{internal}
