% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_utils.R
\name{has_non_genomic_condition}
\alias{has_non_genomic_condition}
\title{Check if a filter condition contains non-genomic constraints}
\usage{
has_non_genomic_condition(fcondition)
}
\arguments{
\item{fcondition}{A filter_condition object}
}
\value{
Logical indicating whether non-genomic conditions exist
}
\description{
Non-genomic conditions are those that don't involve chr or pos.
}
\keyword{internal}
