% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_build.R
\name{infer_build_from_file}
\alias{infer_build_from_file}
\title{Infer genome build from a file}
\usage{
infer_build_from_file(finterface, nsnps = 10000, ...)
}
\arguments{
\item{finterface}{A file_interface object}

\item{nsnps}{Number of SNPs to read for inference}

\item{...}{Additional arguments passed to infer_build}
}
\value{
The inferred genome build
}
\description{
Reads the first nsnps rows from a file and infers the genome build.
}
\keyword{internal}
