% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_class.R
\name{is.filter_condition}
\alias{is.filter_condition}
\title{Test if an object is a filter_condition}
\usage{
is.filter_condition(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether x inherits from filter_condition
}
\description{
Test if an object is a filter_condition
}
\keyword{internal}
