% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_class.R
\name{is_composite_filter_condition}
\alias{is_composite_filter_condition}
\title{Check if a filter condition is composite}
\usage{
is_composite_filter_condition(
  fcondition,
  finterface = get_file_interface(fcondition),
  composite_operators = unname(c(names(composite_fc_operators),
    unlist(composite_fc_operators)))
)
}
\arguments{
\item{fcondition}{A filter_condition object}

\item{finterface}{file_interface for column name resolution}

\item{composite_operators}{Character vector of composite operator names}
}
\value{
Logical indicating whether the condition is composite
}
\description{
Composite conditions combine other conditions with &, |, or parentheses.
}
\keyword{internal}
