% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_interface.R
\name{is_file_interface}
\alias{is_file_interface}
\title{Check if object is a file interface}
\usage{
is_file_interface(finterface)
}
\arguments{
\item{finterface}{Object to test.}
}
\value{
Logical. TRUE if object is a file_interface, FALSE
  otherwise.
}
\description{
Tests whether an object inherits from the "file_interface" class.
}
