% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{is_full_genome}
\alias{is_full_genome}
\title{Check if genomic_regions represents full genome}
\usage{
is_full_genome(gregions)
}
\arguments{
\item{gregions}{A genomic_regions object}
}
\value{
TRUE if gregions has no genomic constraints
}
\description{
A full genome query has all NA values (no constraints).
}
\keyword{internal}
