% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_positions.R
\name{is_genomic_positions}
\alias{is_genomic_positions}
\title{Test if an object is a genomic_positions}
\usage{
is_genomic_positions(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether x inherits from genomic_positions
}
\description{
Test if an object is a genomic_positions
}
\keyword{internal}
