% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_class.R
\name{is_genomic_regions}
\alias{is_genomic_regions}
\title{Check if object is genomic_regions}
\usage{
is_genomic_regions(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
TRUE if x inherits from genomic_regions
}
\description{
Check if object is genomic_regions
}
\keyword{internal}
