% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_utils.R
\name{is_genomic_symbol}
\alias{is_genomic_symbol}
\title{Check if a symbol is a genomic column reference}
\usage{
is_genomic_symbol(symbol)
}
\arguments{
\item{symbol}{A symbol to check}
}
\value{
Logical indicating whether the symbol is chr or pos
}
\description{
Check if a symbol is a genomic column reference
}
\keyword{internal}
