% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_interface.R
\name{is_gzipped}
\alias{is_gzipped}
\title{Check if a file is gzip-compressed}
\usage{
is_gzipped(filename)
}
\arguments{
\item{filename}{Path to the file}
}
\value{
Logical indicating whether the file has a .gz extension
}
\description{
Check if a file is gzip-compressed
}
\keyword{internal}
