% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieugwas_parsing.R
\name{is_ieugwas_file}
\alias{is_ieugwas_file}
\title{Check if a file interface points to an IEUGWAS file}
\usage{
is_ieugwas_file(finterface)
}
\arguments{
\item{finterface}{A file_interface object}
}
\value{
Logical indicating whether the file is IEUGWAS format
}
\description{
IEUGWAS files have 10 columns with specific names and format.
}
\keyword{internal}
