% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_codegen.R
\name{is_simple_read}
\alias{is_simple_read}
\title{Check if this is a simple read with no filtering/processing}
\usage{
is_simple_read(
  branches,
  array_loads,
  column_arrays_before_conditions,
  column_arrays_after_conditions
)
}
\arguments{
\item{branches}{Branch data.table from fcondition_and_rsid_to_awk (or NULL)}

\item{array_loads}{Array-load data.table (may have zero rows)}

\item{column_arrays_before_conditions}{Awk code for pre-condition splits}

\item{column_arrays_after_conditions}{Awk code for post-condition processing}
}
\value{
TRUE if no filtering needed, FALSE otherwise
}
\description{
Check if this is a simple read with no filtering/processing
}
\keyword{internal}
