% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{liftover.genomic_regions}
\alias{liftover.genomic_regions}
\title{Liftover genomic_regions to a different genome build}
\usage{
\method{liftover}{genomic_regions}(x, target, ...)
}
\arguments{
\item{x}{A genomic_regions object}

\item{target}{Target genome build (e.g., "hg38") or chain data.table}

\item{...}{Ignored}
}
\value{
genomic_regions with coordinates in target build
}
\description{
Converts genomic coordinates from one reference genome to another
using UCSC chain files. Handles regions that span chain boundaries
by splitting them.
}
