% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{liftover}
\alias{liftover}
\title{Convert genomic coordinates between genome builds}
\usage{
liftover(x, ...)
}
\arguments{
\item{x}{Object containing genomic coordinates (e.g., genomic_regions)}

\item{...}{Additional arguments passed to methods}
}
\value{
Object with coordinates converted to the target build
}
\description{
S3 generic for lifting over genomic coordinates from one genome build
to another using UCSC chain files.
}
