% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_matching.R
\name{lp_filter_condition}
\alias{lp_filter_condition}
\title{Wrap condition in parentheses}
\usage{
lp_filter_condition(fcondition)
}
\arguments{
\item{fcondition}{Condition list with `condition` element}
}
\value{
Same list with condition wrapped in parentheses
}
\description{
Adds parentheses around a condition for correct operator precedence.
}
\keyword{internal}
