% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_internal.R
\name{lp_wrap_fcondition}
\alias{lp_wrap_fcondition}
\title{Wrap filter condition in parentheses}
\usage{
lp_wrap_fcondition(fcall)
}
\arguments{
\item{fcall}{An existing filter_condition object}
}
\value{
New filter_condition wrapped in lp_filter_condition
}
\description{
Creates a new lp_filter_condition wrapping an existing filter_condition.
Preserves build and genomic_regions attributes.
}
\keyword{internal}
