% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_conversion.R
\name{lt_filter_condition}
\alias{lt_filter_condition}
\title{Build awk less-than condition}
\usage{
lt_filter_condition(var1, var2)
}
\arguments{
\item{var1}{Left operand (column ref or value)}

\item{var2}{Right operand (column ref or value)}
}
\value{
List with `condition` element: "var1 < var2"
}
\description{
Build awk less-than condition
}
\keyword{internal}
