% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_utils.R
\name{make_pos_region}
\alias{make_pos_region}
\title{Create a genomic region data.table from a position boundary}
\usage{
make_pos_region(field, value)
}
\arguments{
\item{field}{Either "start" or "end"}

\item{value}{The position value}
}
\value{
A data.table with the specified boundary
}
\description{
Create a genomic region data.table from a position boundary
}
\keyword{internal}
