% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_info.R
\name{match_a1_a2_to_ref}
\alias{match_a1_a2_to_ref}
\title{Generate awk code for allele matching}
\usage{
match_a1_a2_to_ref(a1_bash_index, a2_bash_index, alt_bash_index)
}
\arguments{
\item{a1_bash_index}{Awk column ref for allele1}

\item{a2_bash_index}{Awk column ref for allele2}

\item{alt_bash_index}{Awk column ref for alt (effect allele)}
}
\value{
Awk code string for allele matching
}
\description{
Creates awk if/else block that determines the non-effect allele (nea)
by comparing effect allele (alt) to allele1/allele2.
}
\keyword{internal}
