% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_matching.R
\name{md5_matches}
\alias{md5_matches}
\title{Verify MD5 checksum of a file}
\usage{
md5_matches(file_path, md5_path = paste0(file_path, ".md5"))
}
\arguments{
\item{file_path}{Path to file to verify}

\item{md5_path}{Path to MD5 checksum file}
}
\description{
Verify MD5 checksum of a file
}
\keyword{internal}
