% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{|.genomic_regions}
\alias{|.genomic_regions}
\title{Union operator for genomic_regions}
\usage{
\method{|}{genomic_regions}(e1, e2)
}
\arguments{
\item{e1, e2}{genomic_regions objects}
}
\value{
Combined genomic_regions
}
\description{
Alias for `+.genomic_regions` - combines regions from both sets.
When used as filter conditions, OR means "match any of these regions".
}
