% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_conversion.R
\name{or_filter_condition}
\alias{or_filter_condition}
\title{Combine two conditions with OR (||)}
\usage{
or_filter_condition(fcondition1, fcondition2)
}
\arguments{
\item{fcondition1}{First condition list}

\item{fcondition2}{Second condition list}
}
\value{
Combined condition list with merged arrays/files
}
\description{
Combine two conditions with OR (||)
}
\keyword{internal}
