% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{parse_chain_file}
\alias{parse_chain_file}
\title{Parse a UCSC chain file into a data.table}
\usage{
parse_chain_file(path)
}
\arguments{
\item{path}{Path to the chain file}
}
\value{
data.table with columns: chr, start, end, width, new_chr, offset, rev
}
\description{
Parses chain files following the UCSC format specification.
See https://genome.ucsc.edu/goldenPath/help/chain.html
}
\details{
Chain files map FROM query TO target. For hg19ToHg38.over.chain:
- query = hg19 (source, what we lift FROM)
- target = hg38 (destination, what we lift TO)

Output structure matches what the liftover function expects:
- chr: source chromosome (qName)
- start/end: ranges in source coordinates
- new_chr: target chromosome (tName)
- offset: source_pos - target_pos (so target = source - offset)
- rev: TRUE if query strand is reversed
}
\keyword{internal}
