% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{+.genomic_regions}
\alias{+.genomic_regions}
\title{Union of genomic_regions}
\usage{
\method{+}{genomic_regions}(e1, e2)
}
\arguments{
\item{e1, e2}{genomic_regions objects}
}
\value{
Combined genomic_regions with merged overlaps
}
\description{
Combines two sets of genomic regions, merging overlapping/contiguous
regions.
}
