% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_ops.R
\name{rbind.genomic_regions}
\alias{rbind.genomic_regions}
\title{Combine genomic_regions by rows}
\usage{
\method{rbind}{genomic_regions}(...)
}
\arguments{
\item{...}{genomic_regions objects to combine}
}
\value{
Combined genomic_regions
}
\description{
Combine genomic_regions by rows
}
