% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_interface.R
\name{read_file_interface}
\alias{read_file_interface}
\title{Load a full file from the file_interface}
\usage{
read_file_interface(file_interface)
}
\arguments{
\item{x}{A file_interface object.}
}
\value{
A data.table with the loaded data.
}
\description{
Loads all data from the file_interface without filtering but still providing
post-processing (standardizing names, etc.). RSID-indexed files are not
re-indexed with chr or pos.
}
