% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_class.R
\name{scaffold_coordinate_columns}
\alias{scaffold_coordinate_columns}
\title{Add empty type-correct coordinate columns to a 0-row table}
\usage{
scaffold_coordinate_columns(x, cols)
}
\description{
Add empty type-correct coordinate columns to a 0-row table
}
\keyword{internal}
