% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_chain_path}
\alias{set_chain_path}
\title{Set chain files path}
\usage{
set_chain_path(path = tools::R_user_dir("filterRead", "data"))
}
\arguments{
\item{path}{Path to directory containing UCSC chain files}
}
\description{
Set chain files path
}
