% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_dbsnp_path}
\alias{set_dbsnp_path}
\title{Set dbSNP reference file path}
\usage{
set_dbsnp_path(path = tools::R_user_dir("filterRead", "data"))
}
\arguments{
\item{path}{Path to dbSNP reference files directory}
}
\description{
Set dbSNP reference file path
}
