% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_class.R
\name{set_variants_class}
\alias{set_variants_class}
\title{(Re)apply the `variants` class to a data.table in place}
\usage{
set_variants_class(x)
}
\arguments{
\item{x}{A data.table.}
}
\value{
`x` with class `c("variants", "data.table", "data.frame")`.
}
\description{
(Re)apply the `variants` class to a data.table in place
}
\keyword{internal}
