% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_arrays.R
\name{setup_column_array}
\alias{setup_column_array}
\title{Generate awk code to split and recombine a delimited column}
\usage{
setup_column_array(bash_index, array_name, delimiter = " ", array_length = 1)
}
\arguments{
\item{bash_index}{The awk column reference (e.g., "$1", "$2")}

\item{array_name}{Name for the awk array to store split values}

\item{delimiter}{String used to split the column (default: " ")}

\item{array_length}{Number of elements expected in the split array}
}
\value{
Character string containing awk code for split and recombine
}
\description{
Creates awk code that: (1) splits a column by delimiter into an array,
and (2) recombines the array elements for output using OFS as separator.
Used for columns that contain multiple encoded values.
}
\keyword{internal}
