% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_matching.R
\name{setup_dbsnp}
\alias{setup_dbsnp}
\title{Download dbSNP reference files to the path defined by get_dbsnp_path()}
\usage{
setup_dbsnp(
  build = c("b37", "b38"),
  type = c("common", "all"),
  max_tries = 3,
  timeout = 3600
)
}
\arguments{
\item{build}{Character vector of builds to download
(default: c("b37", "b38"))}

\item{type}{Type of SNP reference set (common = .01 <= MAF)
("common" or "all")}

\item{max_tries}{Maximum number of retry attempts per file
(default: 3)}

\item{timeout}{Timeout in seconds for each download (default: 3600)}
}
\description{
Download dbSNP reference files to the path defined by get_dbsnp_path()
}
