% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_utils.R
\name{split_genomic_conditions}
\alias{split_genomic_conditions}
\title{Split genomic conditions from a filter condition}
\usage{
split_genomic_conditions(fcondition, build = NULL)
}
\arguments{
\item{fcondition}{A filter_condition object}

\item{build}{Optional genome build}
}
\value{
Modified filter_condition with genomic conditions extracted
}
\description{
Separates genomic (chr, pos) conditions into the genomic_regions attribute
while keeping non-genomic conditions in the filter expression.
}
\keyword{internal}
