% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_condition_utils.R
\name{strip_parentheses}
\alias{strip_parentheses}
\title{Remove parenthesis wrappers from a filter condition}
\usage{
strip_parentheses(fcondition, recursive = TRUE)
}
\arguments{
\item{fcondition}{A filter_condition object}

\item{recursive}{Logical, whether to recursively strip from nested conditions}
}
\value{
Filter condition with parentheses removed
}
\description{
Unwraps lp_filter_condition (parenthesis) nodes from a filter condition.
This always removes all parentheses at the highest level.
If recursive = TRUE, parentheses within the filter_condition are removed as
well.
}
\keyword{internal}
