% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_class.R
\name{validate_genomic_regions_dt}
\alias{validate_genomic_regions_dt}
\title{Validate a genomic_regions data.table}
\usage{
validate_genomic_regions_dt(genomic_regions_dt)
}
\arguments{
\item{genomic_regions_dt}{Object to validate}
}
\value{
The input invisibly; stops with error if invalid
}
\description{
Checks that the object is a data.table.
}
\keyword{internal}
