% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_filter_condition.R
\name{variant_key_expr}
\alias{variant_key_expr}
\title{Awk expression that builds a row's composite variant key}
\usage{
variant_key_expr(finterface)
}
\arguments{
\item{finterface}{A file_interface object.}
}
\value{
A string like `$1 "_" $2 "_" $3 "_" $4`.
}
\description{
Awk expression that builds a row's composite variant key
}
\keyword{internal}
