% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_codegen.R
\name{wrap_filename}
\alias{wrap_filename}
\title{Wrap filename for awk input}
\usage{
wrap_filename(finterface)
}
\arguments{
\item{finterface}{File interface with filename and gzipped flag}
}
\value{
Filename or process substitution string
}
\description{
For gzipped files, wraps in process substitution: <(zcat file.gz)
For regular files, returns filename as-is.
}
\keyword{internal}
