% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_matching.R
\name{add_allele_matching_to_column_info}
\alias{add_allele_matching_to_column_info}
\title{Add information to column_info for allele matching}
\usage{
add_allele_matching_to_column_info(column_info)
}
\arguments{
\item{column_info}{data.table with column metadata}

\item{return}{data.table with modified data (invisible).}
}
\description{
Modify the columns of column_info to include the relevant information for
awk-based matching of allele1/allele2 as effect allele and non-effect
allele. This is done by reference and column_info is returned invisibly.
}
\keyword{internal}
