% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_convert.R
\name{as_genomic_regions}
\alias{as_genomic_regions}
\alias{as_genomic_regions.data.frame}
\title{Convert objects to genomic_regions}
\usage{
as_genomic_regions(x, build = get_build(x), ...)

\method{as_genomic_regions}{data.frame}(x, build = get_build(x), ...)
}
\arguments{
\item{x}{Object to convert}

\item{build}{Genome build (e.g., "hg19", "hg38")}

\item{...}{Additional arguments passed to methods}
}
\value{
A genomic_regions object
}
\description{
Generic function to convert various objects (data.frames, filter_conditions)
to genomic_regions format.
}
\keyword{internal}
