% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{as_genomic_regions}
\alias{as_genomic_regions}
\alias{as_genomic_regions.default}
\title{Coerce object to genomic_regions}
\usage{
as_genomic_regions(x, ...)

\method{as_genomic_regions}{default}(x, build = get_build(x), ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Additional arguments passed to methods}
}
\value{
A genomic_regions object
}
\description{
Coerce object to genomic_regions
}
