% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_class.R
\name{as_variants}
\alias{as_variants}
\alias{as_variants.variants}
\alias{as_variants.data.table}
\title{Coerce an object to variants}
\usage{
as_variants(x, build = build(x), ...)

\method{as_variants}{variants}(x, build = build(x), ...)

\method{as_variants}{data.table}(x, build = build(x), ...)
}
\arguments{
\item{x}{Object to coerce (currently a data.table).}

\item{build}{Genome build; defaults to `build(x)`.}

\item{...}{Unused.}
}
\value{
A `variants` object.
}
\description{
Coerce an object to variants
}
