% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_arrays.R
\name{awk_split_column}
\alias{awk_split_column}
\title{Generate awk split() function call}
\usage{
awk_split_column(bash_index, array_name, delimiter)
}
\arguments{
\item{bash_index}{The awk column reference (e.g., "$1", "$2")}

\item{array_name}{Name for the awk array to store split values}

\item{delimiter}{String used to split the column}
}
\value{
Character string like `split($1, arr, ";")`
}
\description{
Creates awk code to split a column value into an array using a delimiter.
}
\keyword{internal}
