% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{build}
\alias{build}
\alias{build.default}
\alias{build.filter_condition}
\alias{build.character}
\alias{build.list}
\alias{build.file_interface}
\title{Get genome build from an object}
\usage{
build(x)

\method{build}{default}(x)

\method{build}{filter_condition}(x)

\method{build}{character}(x)

\method{build}{list}(x)

\method{build}{file_interface}(x)
}
\arguments{
\item{x}{Object to get build from}
}
\value{
Build string or NULL if not set
}
\description{
S3 generic for extracting the genome build (e.g., "b37", "b38") from
various object types used in filterRead.
}
