% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_class.R
\name{drop_chr_prefix}
\alias{drop_chr_prefix}
\title{Remove prefix from chromosome names}
\usage{
drop_chr_prefix(chr, chr_prefix = "chr")
}
\arguments{
\item{chr}{Character vector of chromosome names}

\item{chr_prefix}{Prefix to remove (defaul: "chr")}
}
\value{
Chromosome names without prefix
}
\description{
Remove prefix from chromosome names
}
\keyword{internal}
