% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions_class.R
\name{full_genomic_regions}
\alias{full_genomic_regions}
\title{Create full-genome genomic_regions (all positions)}
\usage{
full_genomic_regions(build = NULL)
}
\arguments{
\item{build}{Genome build identifier}
}
\value{
genomic_regions representing the entire genome
}
\description{
Returns a genomic_regions with NA for chr, start, and end.
}
\keyword{internal}
