% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk.R
\name{increase_indent}
\alias{increase_indent}
\title{Add indentation to multiline code}
\usage{
increase_indent(code)
}
\arguments{
\item{code}{Awk code string (may contain newlines)}
}
\value{
Code with each line indented by 2 spaces
}
\description{
Add indentation to multiline code
}
\keyword{internal}
