% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_matching.R
\name{needs_a1_a2_to_ref_matching}
\alias{needs_a1_a2_to_ref_matching}
\title{Check if allele matching is needed}
\usage{
needs_a1_a2_to_ref_matching(column_info)
}
\arguments{
\item{column_info}{data.table with column metadata}
}
\value{
TRUE if file has allele1, allele2, alt but no ref
}
\description{
Check if allele matching is needed
}
\keyword{internal}
