% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_matching.R
\name{or_filter_condition_rsid}
\alias{or_filter_condition_rsid}
\title{Combine RSID filter conditions with OR}
\usage{
or_filter_condition_rsid(fcondition1, fcondition2)
}
\arguments{
\item{fcondition1}{First filter condition}

\item{fcondition2}{Second filter condition}
}
\value{
Combined condition list
}
\description{
Combine RSID filter conditions with OR
}
\keyword{internal}
