% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{replace_sentinel_with_na}
\alias{replace_sentinel_with_na}
\title{Restore NA values from sentinel values after foverlaps}
\usage{
replace_sentinel_with_na(
  genomic_regions,
  chr_sentinel = "",
  start_sentinel = 0L,
  end_sentinel = .Machine$integer.max - 1L
)
}
\arguments{
\item{genomic_regions}{data.table with sentinel values}

\item{chr_sentinel}{Value used for NA chromosomes}

\item{start_sentinel}{Value used for NA start}

\item{end_sentinel}{Value used for NA end}
}
\value{
Copy with sentinels replaced by NAs
}
\description{
Converts sentinel values back to NA after overlap operations.
}
\keyword{internal}
