% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_regions.R
\name{validate_genomic_regions_dt}
\alias{validate_genomic_regions_dt}
\title{Validate genomic_regions data.table structure}
\usage{
validate_genomic_regions_dt(genomic_regions_dt)
}
\arguments{
\item{genomic_regions_dt}{data.table to validate}
}
\value{
Invisible input (for chaining)
}
\description{
Validate genomic_regions data.table structure
}
\keyword{internal}
